[nrfconnect] Moved NFC mgr impl from Zephyr to nrfconnect
The NFCOnboardingPayloadManagerImpl that is currently located
in Zephyr is actually nrfconnect specific implementation,
as it includes nrf internal headers that are not available
for any other platform. It seems it was accidentally added
to Zephyr in the past, so it should be moved to nrfconnect.
